% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaHelixCalculator.R
\name{alphaHelixCalculator}
\alias{alphaHelixCalculator}
\title{alphaHelixCalculator}
\usage{
alphaHelixCalculator(
  pathFileInput = "C:/Users/Shashank/Desktop/peptides_second rep.csv",
  pathworkingDir = "C:/Users/Shashank/Desktop",
  pathFolderOutput = NULL
)
}
\arguments{
\item{pathFileInput}{file name from which bla bla bla.}

\item{pathFolderOutput}{file name from which bla bla bla.}
}
\value{
bla bla bla
}
\description{
This function does bla bla bla.
}
\details{
DETAILS
}
\examples{
\dontrun{
alphaHelixCalculator("someFileName")
}
}
