% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSheetCalculation.R
\name{betaSheetCalculation}
\alias{betaSheetCalculation}
\title{betaSheetCalculation}
\usage{
betaSheetCalculation(df, sampleNames, sampleNamesUpdate, dateTimeCurrent)
}
\arguments{
\item{df}{Dataframe for input file data.}

\item{sampleNames}{Names of the samples found in the input file.}

\item{sampleNamesUpdate}{Updated names of the samples found in the input file.}

\item{dateTimeCurrent}{Date and time at the time the simulation began.}
}
\description{
betaSheetCalculation
}
