% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSheetCalculator.R
\name{betaSheetCalculator}
\alias{betaSheetCalculator}
\title{betaSheetCalculator xxx xxx xxx}
\usage{
betaSheetCalculator(
  pathFileInput = "C:/Users/Shashank/Desktop/peptides_second rep.csv",
  pathDirOutput = "C:/Users/Shashank/Downloads"
)
}
\arguments{
\item{pathFileInput}{input file path from which bla bla bla xxxx xxxx xxxx xxxx xxxx xxxx xxxx.}

\item{pathDirOutput}{directory path to which the output files will be generated xxxx xxxx xxxx.}
}
\value{
xxxx xxxx xxxx xxxx xxxx xxxx xxxx xxxx
}
\description{
xxxx xxxx This function does bla bla bla xxxx xxxx.
  xxxxx xxxxxx xxxxx xxxx
  xxxxx xxxxxx xxxxx xxxx
  xxxxx xxxxxx xxxxx xxxx
  xxxxx xxxxxx xxxxx xxxx
}
\details{
DETAILS
}
\examples{
\dontrun{
betaSheetCalculator( pathFileInput    = "<someInputFileName>",
                      pathDirOutput = "<someOutputFolderName>")
}
}
