% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondaryStructureCalculator.R
\name{secondaryStructureCalculator}
\alias{secondaryStructureCalculator}
\title{secondaryStructureCalculator xxx xxx xxx}
\usage{
secondaryStructureCalculator(
  pathFileInput = "C:/Users/Shashank/Desktop/peptides_second rep.csv",
  pathDirOutput = "C:/Users/Shashank/Downloads"
)
}
\arguments{
\item{pathFileInput}{input file path from which bla bla bla xxxx xxxx xxxx xxxx xxxx xxxx xxxx.}

\item{pathDirOutput}{directory path to which the output files will be generated xxxx xxxx xxxx.}
}
\value{
xxxx xxxx xxxx xxxx xxxx xxxx xxxx xxxx
}
\description{
xxxx xxxx This function does bla bla bla xxxx xxxx.
  xxxxx xxxxxx xxxxx xxxx
  xxxxx xxxxxx xxxxx xxxx
  xxxxx xxxxxx xxxxx xxxx
  xxxxx xxxxxx xxxxx xxxx
}
\details{
DETAILS
}
\examples{
\dontrun{
secondaryStructureCalculator( pathFileInput    = "<someInputFileName>",
                      pathDirOutput = "<someOutputFolderName>")
}
}
